requirements.txt
# Core dependencies
GitPython>=3.1.30
PyYAML>=6.0
python-gitlab>=3.13.0
requests>=2.28.2
typing-extensions>=4.5.0

# Code analysis
pylint>=2.17.0
radon>=5.1.0
bandit>=1.7.5
astroid>=2.15.0

# Data processing and visualization
pandas>=1.5.3
matplotlib>=3.7.1
tabulate>=0.9.0
jinja2>=3.1.2

# Security scanning
safety>=2.3.5
detect-secrets>=1.4.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# CLI
click>=8.1.3
rich>=13.3.4
tqdm>=4.65.0

# Documentation
sphinx>=6.1.3
